requirements.txt
Flask~=3.0.3
flask_session~=0.8.0
mysql-connector-python~=9.1.0
redis~=5.2.1
colorama~=0.4.6
pyjwt~=2.10.1
python-dotenv~=1.0.1
gunicorn~=23.0.0